secure api

+++ src/api/articles.js
@@ -5,6 +5,19 @@ const app       = require("../app.js");
 app.route("/api/articles/create")
     .get((req, res) => res.status(503).send({ status: "ERROR"}))
     .post((req, res) => {
+        if (typeof req.body.title !== "string" || req.body.title === "") {
+            res.status(503).send({ status: "Error", extra: "Vous devez rensigner un titre" });
+            return;
+        }
+        if (typeof req.body.content !== "string" || req.body.content === "") {
+            res.status(503).send({ status: "Error", extra: "Le contenu de l'article est vide" });
+            return;
+        }
+        if (typeof req.body.author !== "string" || req.body.author === "") {
+            res.status(503).send({ status: "Error", extra: "L'auteur n'est pas reinsigné"  });
+            return;
+        }
+
         const sqlConnection = mysql.createConnection(sqlConfig);
         sqlConnection.query(
             "INSERT INTO node_articles(title, content, author) VALUES (?, ?, ?);",
@@ -24,6 +37,11 @@ app.route("/api/articles/create")
 app.route("/api/articles/delete")
     .get((req, res) => res.status(503).send({ status: "ERROR"}))
     .post((req, res) => {
+        if (typeof req.body.id !== "string" || req.body.id === "") {
+            res.status(503).send({ status: "Error", extra: "Vous devez rensigner ue Id" });
+            return;
+        }
+        
         const sqlConnection = mysql.createConnection(sqlConfig);
         sqlConnection.query(
             "DELETE FROM node_articles WHERE id = ?;",
@@ -34,7 +52,11 @@ app.route("/api/articles/delete")
                     res.status(503).send({ status: "ERROR" });
                 } else {
                     console.log(result);
-                    res.send({ status: "OK" });
+                    if (result.affectedRows === 0) {
+                        res.status(503).send({ status: "ERROR" , extra: "l'article n'existe pas." });
+                    } else {
+                        res.send({ status: "OK" });
+                    }
                 }
                 sqlConnection.end();
             }
